backend/requirements.txt
pandas
python-dotenv
google-genai
ipython
fastapi
uvicorn
